Aggiunto supporto al reindirizzamento alla pagina di dettaglio per hotel sul click della RecyclerView della home page

+++ app/src/main/java/com/quiriletelese/troppadvisorproject/models/Accomodation.java
@@ -15,7 +15,7 @@ public abstract class Accomodation implements Serializable {
     protected String id;
     protected String name;
     protected Integer avarageRating;
-    protected Double avaragePrice;
+    protected Integer avaragePrice;
     protected String phoneNumber;
     protected Address address;
     protected Point point;
@@ -49,11 +49,11 @@ public abstract class Accomodation implements Serializable {
         this.avarageRating = avarageRating;
     }
 
-    public Double getAvaragePrice() {
+    public Integer getAvaragePrice() {
         return avaragePrice;
     }
 
-    public void setAvaragePrice(Double avaragePrice) {
+    public void setAvaragePrice(Integer avaragePrice) {
         this.avaragePrice = avaragePrice;
     }
